extracts NetworkConfig typing to own file
